Update patch set 16
Patch Set 16:

(2 comments)

Ok, I looked at the EVT schematics. The plan is to have the accel in the lid powered by pp5000 (as it is on p2b). pp5000 is off in G3, and on in every other power state. So, I think we need to add a flag for whether or not any given motion sensor is powered in G3. We can then add some hooks in motion_sense.c for entering and exiting G3. On entering G3, we should stop sampling any accel that is not powered in G3. On exiting G3, we need to run the accel_init() function and start sampling that accel. So, for samus, the lid is accel is off in G3, and the base accel is on in G3.

Patch-set: 16