pd: handle ACCEPT and SOURCE_CAP after a soft reset
After a soft reset, we have an artificial delay so that we don't talk to
the port partner too soon. However, we need to be able to respond to the
message from the partner. This includes:
  - ACCEPT: Sent after the partner received SOFT_RESET. We do nothing in
            this case, but we should not go to HARD_RESET state, which
	    is what we're currently doing.
  - SOURCE_CAP: If we are sink, we should process it. Otherwise, we may
                miss the first SOURCE_CAP packet from the partner.

Also, to properly handle ACCEPT, a new state SOFT_RESET_SENT is added.
This ensures that we go to HARD_RESET state if ACCEPT is not seen in
time.

BUG=chrome-os-partner:31296
TEST=Along with the next CL to send soft reset on dropped ping, check
power doesn't drop when a ping is dropped.
TEST=Modify Samus to not send ACCEPT when a SOFT_RESET packet is
received. Check that we received HARD_RESET after SOFT_RESET.
BRANCH=None

Change-Id: Iddce33befa0c3c43228e68aac8e481d3da52db2a
Signed-off-by: Vic Yang <victoryang@chromium.org>