pd: ensure names of PD states are up-to-date
As we add more PD states, it's easy to forget to update the names of PD
states. This doesn't break any PD functionality so it would be hard to
discover. However, it can easily confuse us when we are debugging. Add
a compile-time assertion to make sure it's updated.

BUG=None
TEST=Remove one names and check build fails.
BRANCH=None

Change-Id: I8b503e361b3418835cdf510dd39481eb7d998035
Signed-off-by: Vic Yang <victoryang@chromium.org>