Update patch set 7
Patch Set 7:

I believe this is now ready to go.  I have abandoned some of the functionality of the original patch that was targeted towards replacing the existing console driver as that was turning into a lot more work than just getting a solution for forwarding multiple consoles over USB.  So, this code should be looked at as a driver for forwarding consoles.  And a base to work from to eventually replace the existing console USART driver, but that is a secondary priority.  I would like to land this more minimal version and pipe it through USB, then work on replacing the console USART driver.

Patch-set: 7
Label: Code-Review=0