pd: zinger: fix bug in pd transmit retry mechanism
This fixes a bug in the PD transmit retry mechanism.

There were two bugs:
On zinger, the task_wait_event() was returning the wrong mask in the case
of a timeout, which is used during pd transmit retries to distinguish
between timing out waiting for a reply and getting the first edge of a
reply.

In the retry mechanism, we were assuming that only a pd rx interrupt
will wake up this event. But, there are other events that could potentially
wake us up, so we need to check to make sure that pd rx started when
we first wake up.

BUG=chrome-os-partner:30135
BRANCH=none
TEST=load onto samus and zinger. run "pd 0 flash rw_hash" a bunch of
times manually from the console. Observe that we nearly always fail
the first receive, but succeed on next try, which prevents us from
dropping the negotiation.

Change-Id: I5f7261176c151c3185d76aa374b9b83ac9df9a7d
Signed-off-by: Alec Berg <alecaberg@chromium.org>