Update patch set 1
Patch Set 1:

Not sure if this is the best way to accomplish the task, there may be a better solution that is more extensible for more of a host event pass through where the PD can send a host event with a target destination, just like host commands. but, that is really complicated to accomplish one host event.

Patch-set: 1
Label: Code-Review=0