Veyron: Change WARM_RESET gpio to open drain
The servo and the EC both output on the WARM_RESET line. Servo is open drain but
the EC was not. This caused a short whenever the servo tried to assert the
reset. The button still worked because the button is a lot stronger at pulling
down than the servo gpio.

Changing the WARM_RESET EC pin to open drain is an easy fix to this problem
without changing the hardware.

BRANCH=None
BUG=None
TEST=Warm reset via servo command should work. Warm reset button on servo always worked.

Change-Id: Ib615bc438a5726e40b0b502a197a57dbea6ee780
Signed-off-by: Alexandru M Stan <amstan@chromium.org>