Update patch set 2
Patch Set 2: Code-Review-1

This change should not be needed. The noise on AC present is likely a hardware problem on your board. We need to test at least on a couple boards to verify if this is a problem. If it is a problem, we should create a bug specifically for it and investigate electrically first.

Patch-set: 2
Reviewer: Gerrit User 1103909 <1103909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1