Update patch set 1
Patch Set 1:

Sonny: It's up to dexter.

I haven't heard a good way to change the hardware yet, most solutions are pretty complex(a resistor in series will drop the voltage too much to the point where the reset won't be asserted), the nicest hardware way would be to use a fet that pulls the line down(but the open drain GPIO pin mode is exactly that...).

I suggest we leave the hardware as is.

Patch-set: 1