Flashing: Force board into reset for flashing
This ensures that nothing previously flashed to the baord can
interfere with the flashing operation (by wedging the MCU or
putting it into a state that the falshing code can't handle).

This also adds a dependency on ec.bin to the flash
target, ensuring that the firmware image is up to date
when flashing.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=none
TEST=make buildall -j

Change-Id: I8cdfa6f5c84ed84d6b6e6b30d6683a23087f2c63