Update patch set 2
Patch Set 2: Verified+1

This is similar to the MUX_GPIO mechanism, but is somewhat different. We need to gate SCL for any other I2C transaction, so this needs to be done within i2c_xfer (when the i2c port is locked.) Plus that this is a hack that is supposed to be reverted when the chip is fixed, so I didn't want to make the MUX_GPIO mechanism uglier just for this.

Patch-set: 2
Reviewer: Gerrit User 1001139 <1001139@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1