accel: fix bug, initialize accels every time we boot out of G3
This makes sure that we initialize all accelerometers when we leave
G3 because some accelerometers are not powered in G3.

Also, changed some of the print statements in motion_sense.c to
help in debugging.

BUG=none
BRANCH=none
TEST=tested on samus. verified if you go back and forth between
G3 and S0 that the lid accelerometer is always initialized by
using accelread command to verify that data is being updated.

Change-Id: I73effda4e6b04a629851e6c310d53b044c4aad42
Signed-off-by: Alec Berg <alecaberg@chromium.org>