Update patch set 1
Patch Set 1:

I seems that it kinda works..

The watchdog and the UART are broken. They will not wakeup the chip. I have to disable the watchdog for the EC to stop resetting. UART works after a wakeup, but it will not cause a wakeup.

SPI does cause a wakeup! Woo!

Patch-set: 1