Update patch set 2
Patch Set 2: Code-Review-1

Yeah... I probably want to look at this more.

It seems the only reason v3 works is because out_preamble gets partially eaten(and it doesn't matter much in the big picture because the bytes that are eaten are repeated just before the dma anyway).

Patch-set: 2
Label: Code-Review=-1