Update patch set 3
Patch Set 3: Code-Review-1 Verified+1

(2 comments)

There should not be functional changes between this and the previous commit, just a few #ifdefs to isolate the fixes to F0 and a fix for the end of message(which doesn't matter much).

TODO is to fix protocol v2. But if you have a new kernel it should work just fine without it.

Feel free to use this in your EC right now. But do not apply https://chromium-review.googlesource.com/#/c/217201/

Patch-set: 3
Label: Code-Review=-1
Label: Verified=+1