WIP: Debugging STM32F0 Debug RX
Seems that from time to time the EC enters a mode where it doesn't want to talk
to anybody. It persists across ap reboots as well.

This patch adds a lot of debugging features to debug this.

It seems that the RX DMA now gets out of sync(confirmed by the 3 extra 00s at
the start of the message). "unwedgerx 3" doesn't seem to fix it. fifostatus has
no extra rx characters waiting, yet they appear as part of the dma transfer:

	in_msg 0x 00 00 00 03 aa 9e 00 00 00 0b
	[899.665246 SPI rx bad data]
	in_msg 0x 00 00 00 03 aa 9e 00 00 00 0b
	[899.774912 SPI rx bad data]

	> fifostatus
	SPI_sr=0x1840, FTLVL=3 FRLVL=0
	>
	> unwedgerx 3
	Consuming 3 bytes: 00 00 00 done
	> in_msg 0x 00 00 00 03 fb 02 00 00 00 00
	[920.845283 SPI rx bad data]

	>
	> fifostatus
	SPI_sr=0x1840, FTLVL=3 FRLVL=0

I implemented an automatic offset that will ignore the first bunch of 00s.

I'm also resetting the SPI after every CS rising edge. This fixes the DMA
problem in a nice non hacky way (hurray!).

BRANCH=None
BUG=None
TEST=ap# cd /sys/class/power_supply/sbs-20-000b/; while true; do grep "" * >/dev/null 2>&1; done
You should not see errors or rxoffset messages.

Change-Id: If9ab93c5c9040a2c7bda33d7cc990603f1121f3f
Signed-off-by: Alexandru M Stan <amstan@chromium.org>