Update patch set 2
Patch Set 2:

(2 comments)

Thanks for the review, I will change most of these.

Deciding not to make a new file for rtc stuff right now. It fits in well with the clock module, which is the only place we need access to the RTC, and matches other ECs, like stm32f100.

Patch-set: 2