stm32-USB: USB SPI tunnel driver
Simple control of SPI for flashing over USB.

This driver is working, and using the discovery board
with a W25Q16 flash chip attached flashrom can read,
erase, write, and verify the whole chip in 45 seconds.

There is still work to be done around enabling and
disabling this tunnel, in particular, the board init
function is just pulled wholesale from fruitpie and
should be trimmed and removed in favor of configuration
in gpio.inc.  There should also be usb_spi_enable and
usb_spi_disable functions added.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I224f1f87cd6adc8b64c17de1df98dae0a9cfa6a5