Update patch set 1
Patch Set 1: Verified+1

I plan to clean up the initialization of SPI as well as add the enable/disable API before landing this, just wanted you guys to see and comment on this part.  I also wrote a flashrom backend that Dave and Todd are looking at that uses this driver to flash a SPI chip (takes about 44 seconds to read/erase/write/verify a 2MB chip).

Patch-set: 1
Label: Verified=+1