stm32f0: Change uart clock to HSI
When waking up from sleep, the real CPU_CLOCK is a lie for a moment(since we
cannot switch to the real clock during the first character) so the first
character will be corrupted.

The UART clock is now sourced from HSI(8MHz) which is available from the first
moment after the cpu wakes up from sleep.

BUG=None
TEST=Console should work.
When waking up(not implemented yet) it will also not lose a character
BRANCH=None

Change-Id: Ia12ed0634290f3edadfe3471b311759c3176260e
Signed-off-by: Alexandru M Stan <amstan@chromium.org>