Update patch set 1
Patch Set 1: Code-Review-2 Verified-1

Not ready to submit. Need to disable deep sleep when USB PD is active, and need to figure out what to do with i2c slave, as it seems to be causing the EC intermittent problems when PD is in deep sleep.

Patch-set: 1
Label: Code-Review=-2
Label: Verified=-1