stm32f0: samus_pd: console uart clock source reset on sysjump
Add code to reset the console UART clock source to default on
a sysjump. This fixes a bug in a recent CL where the console
UART clock source was changed, but then a sysjump to an earlier
version would break the console.

BUG=chrome-os-partner:32143
BRANCH=none
TEST=load onto samus, allow a software sync to an older version
and verify that the console works after sysjump between RO and RW.

Change-Id: If8a145ce4206017d7ea4c3915c3313d14e8a2924
Signed-off-by: Alec Berg <alecaberg@chromium.org>