Update patch set 1
Patch Set 1: Code-Review+1

I was trying to figure out how this works in charge_state_v1, and I think when AC interrupt occurs it wakes the charger task as well as sending a host event. So, by the time the AP sends host command, the charger task will hopefully have run already and updated the battery flags. I think this solution is more robust, but do we have to have another host command? Can we just re-use EC_HOST_EVENT_BATTERY and combine the if() check with the capacity changed if() statement?

Patch-set: 1