Update patch set 1
Patch Set 1: Code-Review-1

I feel like this might need to be moved to the board/veyron/ folder as it does not relate to rockchip based designs. If we make another board based on rockchip we might not use this regulator hookup. Randall, what do you think?

Another thing is that this is related to VCC5V_GOOD_L, this CL should probably check that the supply actually went up.

Patch-set: 1