Update patch set 1
Patch Set 1:

If we're really hurting for space, we can save another half the table by abusing the fact that for pi/4 <= x <= pi/2, sin(x) = 1 - sin(pi/2 - x), so instead of storing half the sine table, we only really need to store a quarter of it.

I'm not sure if we're worried about packing a lot of logic into cycle_010 or not, otherwise that'd be a no-brainer.

Patch-set: 1
Label: Code-Review=0