Update patch set 1
Patch Set 1:

I know this is still sorta hacky. It'd be nice to fix this by allowing to change individual i2c busses into normal mode, but that would require a significant amount of refactoring. I'll keep the original bug open though so when we get around to doing more cleanup, this can be fixed better.

Patch-set: 1