Pinky: Trying to fix hibernate wakeup
Power button was disabled for now, it requires it to be inverted.
This patch also requires a few resistors changed on the remaining wake sources.

I changed the default state of the LEDs to be off, so the watchdog wakeup from
hibernate will not blink the leds anymore.

BRANCH=chrome-os-partner:32782
BUG=None
TEST=type "hibernate" on a modified proto2

Change-Id: I34306e37530517b13123f44ebabbfa0235f7782f
Signed-off-by: Alexandru M Stan <amstan@chromium.org>