lightbar: converting sequences to hardcoded programs
This pays for the interpreter expansion and saves another 500
bytes on top of that. We get rid of a lot of code this way
without changing the API, since there is still cleanup to do.

When we convert everything to use programs and shove them into
the kernel, we'll be able to remove all these hardcoded programs,
and save another 200+ bytes right away.

BUG=chrome-os-partner:32203
BRANCH=ToT
TEST=Run on hw and simulator, verify there are no differences.

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Change-Id: Ibc65c6c1f7f5561b3774cf2d0a89c1665345b0d0