samus: Change 3.3DSW_GATED to enabled in S0 and disabled in S3
Move the 3.3DSW_GATED rail to be enabled in transition to S0 and
disabled in the transition to S3.

This is the rail for the core regulator and temp sensors so it
does not need to be enabled in S5.

BUG=chrome-os-partner:32382
BRANCH=samus
TEST=build and boot on samus, successfully suspend/resume and
ensure that PP3300_DSW_GATED is turned off in S3.

Change-Id: Ic47f81860e3f0cb7b5d81ba96181b7ee7cf72f66
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>