Fix incorrect valid and writable flash flags
The valid and writable flags the EC sends back to the AP are incorrect.
They are a little bit different on differnt chips, so let's move it to
flash physical layer. This is not any causing problem, but we should fix
this.

BUG=chrome-os-partner:32745
TEST=make buildall
BRANCH=samus

Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222661
Reviewed-by: Randall Spangler <rspangler@chromium.org>
(cherry picked from commit f8fd63f135b5922900ceeb345857070e814f8877)
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>

Change-Id: Ibcda5ae770f5ea02cde094490997a5bc447df88f