EC: Ensure the udelay function waits at least the time indicated.
I am working with udelay(4) and sometimes get the tick like,
6->7->7->8->8->9->A and then the udelay return. But, the sampling
point of 6 could be at the end of 6(close to 7) and the point of
A could be right at the beginning of A(close to A). This function
could get delay from (us - 1) to (us + 1). This change is to ensure
the delay at least over the parameter, us.

BRANCH=master
BUG=None
TEST=Build an EC FW iamge and run on Rambi to ensure at the time
duration indicated by the parameter is elaspsed and satisfied.
Signed-off-by: Kenji Chen <kenji.chen@intel.com>
Change-Id: I797f80c577d7e29e75a304aec1e02d2c750f8a23