Calibrated lightbar values for Samus & Ryu
This updates the lightbar params with the values determined in
the optic lab. The full-on primary colors are of uniform
intensity at the brightness level used by Link. The Google colors
provide the closest match to the official color palette (at full
brightness) described at

https://sites.google.com/a/google.com/brandsite/the-colours

BUG=chrome-os-partner:33017
BRANCH=ToT,samus,ryu
TEST=manual

The calibration and testing was done in the optic lab.

You can see the result just by looking at the lightbar. For best
results, set the brightness to the max with

  ectool lightbar brightness ff

Change-Id: I1cb64e3c547844adf733cd46bff1256319b4f612
Signed-off-by: Bill Richardson <wfrichar@chromium.org>