pd: samus: add support for power swap command
Add support for PR_SWAP command as per PD specification.

WIP: This is a work in progress and is missing a few things:
- Still need to add a seperate variable for data role and use that
data role to know when to send VDMs.
- Hard reset needs to be modified to allow for hard reset after
a power swap, which apparently is different than a normal hard
reset in that after the hard reset, the original power source turns
back on.

BUG=chrome-os-partner:28343
BRANCH=samus
TEST=test by connecting two samus' and running 'pd 1 swap power'
from console. verified that both sides switch power roles by
observing console output.

Change-Id: I0e8738b544de9f9a4348250630e67d0fefb4486d
Signed-off-by: Alec Berg <alecaberg@chromium.org>