usb_pd: Negotiate minimum power for unused ports
When a PD charger is detected, gets its max. charge capability and
determine if the port will be selected as our charge port. If so, begin
to charge from the port with a minimal current until negotiation is
completed. If the port will not be our charge port, instead negotiate
for the minimal power mode.

BUG=chrome-os-partner:32003
TEST=Manual on Samus. Plug in Zinger, verify that current limit is
initially set to 500 mA, then switches to 3000 mA shortly after.
BRANCH=Samus

Change-Id: Ifa562b72d763642fc8bd62bc7f5aaa4eda1ef950
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>