Update patch set 2
Patch Set 2: Code-Review-1

yes, this fixes the stack overflow, but then we are still not waking up in time, which is a much more subtle bug that could bite us. we need to know why it's oversleeping.

also, i think the better fix is to increase stack size, or change idle function to take up less stack space so that we can still print some warning.

Patch-set: 2
Label: Code-Review=-1