Update patch set 2
Patch Set 2:

For reference, tTransmit, the "delay between last bit of the message EOP has been received by the physical layer until the first bit of the Preamble of the GoodCRC message has been transmitted by the physical layer" is 195us. For reference, I measured from the final edge of the previous message to the first falling edge of the goodCRC. So, we probably have to add a few us to my measurements. So, with zinger, we are getting very close...

Patch-set: 2