Update patch set 1
Patch Set 1:

That doesn't really address the problem.  This isn't working around the broken stm32f3 pull up problem.  This is dealing with the fact that the PD USB lines are always connected to the RFU or AP USB 2.0 lines, even when debug mode is not enabled.  So we need a way to bring up the PD without initializing the USB peripheral at all, so that it doesn't interfere with these lines.  Thus a new config.

Patch-set: 1