Update patch set 1
Patch Set 1:

> That doesn't really address the problem.
> This isn't working around the broken stm32f3 pull up problem.

I was not referring to the pull-up issue.
I mentioned gpio_config_module to change the pin muxing later.

> This is dealing with the fact that the PD USB lines are always connected to the RFU
> or AP USB 2.0 lines, even when debug mode is not enabled.
>  So we need a way to bring up the PD without initializing the USB peripheral at all,
> so that it doesn't interfere with these lines.  Thus a new config.

On some STM32 (eg STM32L1xx), the USB lines are high impedance when not transmitting as far as I can tell.
is that different on STM32F3 ?

Patch-set: 1